pyproject.toml
[tool.poetry]
name = "scrapper-new"
version = "0.1.0"
description = ""
authors = ["nobinkhan <nazrulkhan.dk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
djangorestframework = "^3.14.0"
pandas = "^2.2.0"
selenium = "^4.17.2"
beautifulsoup4 = "^4.12.3"
django-adminlte-3 = "^0.1.6"
django-cors-headers = "^4.3.1"
whitenoise = "^6.6.0"
pillow = "^10.2.0"
webdriver-manager = "^4.0.1"
icecream = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
